requirements.txt
flask
flask-login
oauthlib
requests
cryptography
dnspython
pymongo
PyPDF2
fitz
tools
frontend
apiclient
oauth2client
google-api-python-client
flask_bcrypt
flask_wtf
Pillow
google_auth_oauthlib